Split passed into 3 nyazy -> std -> llvm w/ print -> llvm
