+++ src/ir/CMakeLists.txt
@@ -2,7 +2,10 @@ target_sources(NyaZyDialect PRIVATE
     NyaZyDialect.cpp
     NyaZyOpsEnums.cpp
     NyaZyOps.cpp
-    lowerToLLVM.cpp
+    lowerNyaZyPrintToPrintf.cpp
+    lowerNyaZyToStd.cpp
+    lowerStdToLLVM.cpp
+    common.cpp
 )
 
 target_link_libraries(NyaZyDialect
