+++ src/ir/lowerNyaZyToStd.cpp
@@ -1,3 +1,4 @@
+#include "ir/common.h"
 #include "ir/NyaZyDialect.h"
 #include "ir/NyaZyOps.h"
 #include "ir/Pass.h"
@@ -29,76 +30,6 @@
 
 namespace {
 
-// mlir::FlatSymbolRefAttr getOrInsertExit(mlir::ModuleOp module) {
-//     auto* context = module.getContext();
-//     const char *exitSymbol = "exit";
-//
-//     if (module.lookupSymbol<mlir::LLVM::LLVMFuncOp>(exitSymbol)) {
-//         return mlir::SymbolRefAttr::get(context, exitSymbol);
-//     }
-//
-//     auto llvmExitType =
-//     mlir::LLVM::LLVMFunctionType::get(mlir::LLVM::LLVMVoidType::get(context),
-//     {mlir::IntegerType::get(context, 64)});
-//
-//     mlir::PatternRewriter rewriter{context};
-//     mlir::PatternRewriter::InsertionGuard guard(rewriter);
-//     rewriter.setInsertionPointToStart(module.getBody());
-//     rewriter.create<mlir::LLVM::LLVMFuncOp>(
-//         module.getLoc(), exitSymbol, llvmExitType);
-//
-//     return mlir::SymbolRefAttr::get(context, exitSymbol);
-// }
-
-// class ConstantOpLowering : public mlir::OpConversionPattern<nyacc::ConstantOp> {
-// public:
-//   explicit ConstantOpLowering(mlir::MLIRContext *context)
-//       : OpConversionPattern(context) {}
-// 
-//   mlir::LogicalResult
-//   matchAndRewrite(nyacc::ConstantOp op, OpAdaptor adaptor [[maybe_unused]],
-//                   mlir::ConversionPatternRewriter &rewriter) const override {
-//     auto constantOp = mlir::cast<nyacc::ConstantOp>(op);
-//     rewriter.replaceOp(op, rewriter.create<mlir::arith::ConstantOp>(
-//                                op->getLoc(), constantOp.getValue()));
-// 
-//     return mlir::success();
-//   }
-// };
-
-// Helper function to get or create a format string global variable
-mlir::Value getOrCreateGlobalString(mlir::Location loc, mlir::ModuleOp module,
-                                    mlir::PatternRewriter &rewriter,
-                                    llvm::StringRef formatStr) {
-  mlir::LLVM::GlobalOp globalStr;
-
-  // Use the format string as the global variable name
-  std::string globalName = "global_str_" + std::to_string(std::hash<std::string>{}(formatStr.str()));
-
-  if (!(globalStr = module.lookupSymbol<mlir::LLVM::GlobalOp>(globalName))) {
-    mlir::OpBuilder::InsertionGuard guard(rewriter);
-    rewriter.setInsertionPointToStart(module.getBody());
-
-    mlir::Type i8Type = mlir::IntegerType::get(rewriter.getContext(), 8);
-    mlir::Type strType = mlir::LLVM::LLVMArrayType::get(i8Type, formatStr.size() + 1);
-
-    globalStr = rewriter.create<mlir::LLVM::GlobalOp>(
-        loc, strType, /*isConstant=*/true, mlir::LLVM::Linkage::Internal,
-        globalName, rewriter.getStringAttr(formatStr.str() + '\0'));
-  }
-
-  // Get a pointer to the first character in the global string
-  mlir::Type i8PtrType = mlir::LLVM::LLVMPointerType::get(rewriter.getContext());
-
-  mlir::Value globalPtr = rewriter.create<mlir::LLVM::AddressOfOp>(loc, globalStr);
-  mlir::Value zero = rewriter.create<mlir::LLVM::ConstantOp>(
-      loc, rewriter.getI64Type(), rewriter.getI64IntegerAttr(0));
-  mlir::Value ptr = rewriter.create<mlir::LLVM::GEPOp>(
-      loc, i8PtrType, globalStr.getType(), globalPtr, mlir::ValueRange{zero, zero});
-
-  return ptr;
-}
-
 struct ConstantOpLowering : public mlir::OpConversionPattern<nyacc::ConstantOp> {
   using OpConversionPattern<nyacc::ConstantOp>::OpConversionPattern;
 
@@ -117,7 +48,7 @@ struct ConstantOpLowering : public mlir::OpConversionPattern<nyacc::ConstantOp> 
 
       // Create a global string constant
       std::string globalName = "str_const_" + std::to_string(reinterpret_cast<uintptr_t>(op.getOperation()));
-      auto ptr = getOrCreateGlobalString(loc, module, rewriter, strAttr.getValue());
+      auto ptr = nyacc::getOrCreateGlobalString(loc, module, rewriter, strAttr.getValue());
 
       rewriter.replaceOp(op, ptr);
     } else {
@@ -212,9 +143,7 @@ struct CmpOpLowering : public mlir::OpConversionPattern<nyacc::CmpOp> {
 		// auto arithPredAttr = mlir::arith::invertPredicate(arithPred);
 		auto loc = op->getLoc();
 		auto arithCmpOp = rewriter.create<mlir::arith::CmpIOp>(loc, arithPred, adaptor.getLhs(), adaptor.getRhs());
-    auto u1Type = rewriter.getIntegerType(1, false);
-    auto unsignedCmpResult = rewriter.create<mlir::arith::BitcastOp>(loc, u1Type, arithCmpOp);
-    rewriter.replaceOp(op, unsignedCmpResult);
+    rewriter.replaceOp(op, arithCmpOp);
 
     return mlir::success();
   }
@@ -255,7 +184,8 @@ struct CastOpLowering : public mlir::OpConversionPattern<nyacc::CastOp> {
 
     const auto int2intExt = [&](mlir::Value from, mlir::IntegerType intType) -> mlir::Value {
       assert(from.getType().getIntOrFloatBitWidth() < intType.getIntOrFloatBitWidth());
-      if (from.getType().isUnsignedInteger()) {
+      // i1 is a special, treated as unsigned integer
+      if (from.getType().isUnsignedInteger() || from.getType().getIntOrFloatBitWidth() == 1) {
         return rewriter.create<mlir::arith::ExtUIOp>(loc, intType, from);
       } else {
         return rewriter.create<mlir::arith::ExtSIOp>(loc, intType, from);
@@ -431,137 +361,30 @@ struct YieldOpLowering : public mlir::OpConversionPattern<nyacc::YieldOp> {
   }
 };
 
-class NyaZyToLLVMPass
-    : public mlir::PassWrapper<NyaZyToLLVMPass,
+class NyaZyToStdPass
+    : public mlir::PassWrapper<NyaZyToStdPass,
                                mlir::OperationPass<mlir::ModuleOp>> {
 public:
-  MLIR_DEFINE_EXPLICIT_INTERNAL_INLINE_TYPE_ID(NyaZyToLLVMPass)
+  MLIR_DEFINE_EXPLICIT_INTERNAL_INLINE_TYPE_ID(NyaZyToStdPass)
   void getDependentDialects(mlir::DialectRegistry &registry) const override {
-    registry.insert<nyacc::NyaZyDialect>();
+    registry.insert<nyacc::NyaZyDialect, mlir::arith::ArithDialect, mlir::scf::SCFDialect, mlir::memref::MemRefDialect, mlir::func::FuncDialect>();
   }
 
 private:
   void runOnOperation() final;
 };
 
-// Helper function to get or insert the 'printf' function declaration
-mlir::LLVM::LLVMFuncOp getOrInsertPrintf(mlir::ModuleOp module, mlir::PatternRewriter &rewriter) {
-  auto printfFunc = module.lookupSymbol<mlir::LLVM::LLVMFuncOp>("printf");
-  if (printfFunc) {
-    return printfFunc;
-  }
-
-  // Create 'printf' function declaration
-  mlir::OpBuilder::InsertionGuard guard(rewriter);
-  rewriter.setInsertionPointToStart(module.getBody());
-
-  mlir::Type i8PtrType = mlir::LLVM::LLVMPointerType::get(rewriter.getContext());
-
-  auto printfType = mlir::LLVM::LLVMFunctionType::get(
-      mlir::IntegerType::get(rewriter.getContext(), 32),
-      {i8PtrType}, /*isVarArg=*/true);
-
-  printfFunc = rewriter.create<mlir::LLVM::LLVMFuncOp>(
-      rewriter.getUnknownLoc(), "printf", printfType);
-
-  return printfFunc;
-}
-
-
-struct PrintIntOpLowering : public mlir::OpConversionPattern<nyacc::PrintOp> {
-  using mlir::OpConversionPattern<nyacc::PrintOp>::OpConversionPattern;
-
-  mlir::LogicalResult matchAndRewrite(
-      nyacc::PrintOp op, OpAdaptor adaptor,
-      mlir::ConversionPatternRewriter &rewriter) const override {
-    // Check if the operand is an integer type
-    auto operandType = adaptor.getOperand().getType();
-    if (!llvm::isa<mlir::IntegerType>(operandType)) {
-      return mlir::failure();
-    }
-
-    mlir::Location loc = op.getLoc();
-    mlir::ModuleOp module = op->getParentOfType<mlir::ModuleOp>();
-
-    // Get or create the printf function declaration
-    auto printfFunc = getOrInsertPrintf(module, rewriter);
-
-    // Get or create the format string for integers
-    auto formatStrPtr = getOrCreateGlobalString(loc, module, rewriter, "%ld\n");
-
-    // Ensure the operand is of LLVM integer type (i64)
-    mlir::Value value = adaptor.getOperand();
-
-    // Create the printf call
-    auto callop = rewriter.create<mlir::LLVM::CallOp>(
-      loc,
-        mlir::TypeRange{mlir::IntegerType::get(rewriter.getContext(), 32)},
-        mlir::SymbolRefAttr::get(printfFunc),
-        mlir::ValueRange{formatStrPtr, value});
-    
-    auto ptrType = mlir::LLVM::LLVMPointerType::get(rewriter.getContext());
-    mlir::Type printfType = mlir::LLVM::LLVMFunctionType::get(
-        mlir::IntegerType::get(rewriter.getContext(), 32),
-        {ptrType}, /*isVarArg=*/true);
-    callop->setAttr("var_callee_type", mlir::TypeAttr::get(printfType));
-
-    rewriter.eraseOp(op);
-
-    return mlir::success();
-  }
-};
-
-struct PrintStringOpLowering : public mlir::OpConversionPattern<nyacc::PrintOp> {
-  using mlir::OpConversionPattern<nyacc::PrintOp>::OpConversionPattern;
-
-  mlir::LogicalResult matchAndRewrite(
-      nyacc::PrintOp op, OpAdaptor adaptor,
-      mlir::ConversionPatternRewriter &rewriter) const override {
-    // Check if the operand is an LLVM pointer to i8 (string)
-    auto operandType = adaptor.getOperand().getType();
-    auto llvmPtrType = llvm::dyn_cast<mlir::LLVM::LLVMPointerType>(operandType);
-    if (!llvmPtrType) {
-      return mlir::failure();
-    }
-
-    mlir::Location loc = op.getLoc();
-    mlir::ModuleOp module = op->getParentOfType<mlir::ModuleOp>();
-
-    // Get or create the printf function declaration
-    auto printfFunc = getOrInsertPrintf(module, rewriter);
-
-    // Get or create the format string for strings
-    auto formatStrPtr = getOrCreateGlobalString(loc, module, rewriter, "%s\n");
-
-    // Create the printf call
-    auto callop = rewriter.create<mlir::LLVM::CallOp>(
-        loc,
-        mlir::TypeRange{mlir::IntegerType::get(rewriter.getContext(), 32)},
-        mlir::SymbolRefAttr::get(printfFunc),
-        mlir::ValueRange{formatStrPtr, adaptor.getOperand()}
-    );
-    auto ptrType = mlir::LLVM::LLVMPointerType::get(rewriter.getContext());
-    mlir::Type printfType = mlir::LLVM::LLVMFunctionType::get(
-        mlir::IntegerType::get(rewriter.getContext(), 32),
-        {ptrType}, /*isVarArg=*/true);
-    callop->setAttr("var_callee_type", mlir::TypeAttr::get(printfType));
-
-    rewriter.eraseOp(op);
-
-    return mlir::success();
-  }
-};
 
 } // namespace
 
-void NyaZyToLLVMPass::runOnOperation() {
+void NyaZyToStdPass::runOnOperation() {
   mlir::ConversionTarget target(getContext());
-  target.addLegalDialect<mlir::BuiltinDialect, mlir::LLVM::LLVMDialect, mlir::scf::SCFDialect, mlir::memref::MemRefDialect>();
+  target.addLegalDialect<mlir::BuiltinDialect, mlir::LLVM::LLVMDialect, mlir::arith::ArithDialect, mlir::scf::SCFDialect, mlir::memref::MemRefDialect, mlir::func::FuncDialect>();
   target.addIllegalDialect<nyacc::NyaZyDialect>();
   target.addLegalOp<nyacc::PrintOp>();
 
   mlir::RewritePatternSet patterns(&getContext());
-  // nyazy -> arith + func
+  // nyazy -> arith + func + scf + memref
   patterns.add<ConstantOpLowering, FuncOpLowering, ReturnOpLowering,
                AddOpLowering, SubOpLowering, MulOpLowering, DivOpLowering, 
                PosOpLowering, NegOpLowering, CmpOpLowering, CastOpLowering,
@@ -570,66 +393,13 @@ void NyaZyToLLVMPass::runOnOperation() {
                >(
       &getContext());
 
-  // * -> llvm
-  mlir::LLVMTypeConverter typeConverter(&getContext());
-  mlir::arith::populateArithToLLVMConversionPatterns(typeConverter, patterns);
-  mlir::populateFuncToLLVMConversionPatterns(typeConverter, patterns);
-  mlir::populateSCFToControlFlowConversionPatterns(patterns);
-  mlir::cf::populateControlFlowToLLVMConversionPatterns(typeConverter, patterns);
-
   if (failed(
           applyFullConversion(getOperation(), target, std::move(patterns)))) {
     signalPassFailure();
   }
 
-  // scf -> llvm
-  mlir::ConversionTarget scfTarget(getContext());
-  mlir::LLVMTypeConverter scfTypeConverter(&getContext());
-  scfTarget.addLegalDialect<mlir::BuiltinDialect, mlir::LLVM::LLVMDialect, mlir::memref::MemRefDialect>();
-  scfTarget.addLegalOp<nyacc::PrintOp>();
-  scfTarget.addIllegalDialect<mlir::scf::SCFDialect>();
-
-  mlir::RewritePatternSet scfPatterns(&getContext());
-  mlir::populateSCFToControlFlowConversionPatterns(scfPatterns);
-  mlir::cf::populateControlFlowToLLVMConversionPatterns(scfTypeConverter, scfPatterns);
-
-  if (failed(
-          applyFullConversion(getOperation(), scfTarget, std::move(scfPatterns)))) {
-    signalPassFailure();
-  }
-
-  // memref -> llvm
-  mlir::ConversionTarget target2(getContext());
-  target2.addLegalDialect<mlir::BuiltinDialect, mlir::LLVM::LLVMDialect>();
-  target2.addLegalOp<nyacc::PrintOp>();
-  target2.addIllegalDialect<mlir::memref::MemRefDialect>();
-
-  mlir::RewritePatternSet patterns2(&getContext());
-
-  mlir::LLVMTypeConverter typeConverter2(&getContext());
-
-  mlir::populateFinalizeMemRefToLLVMConversionPatterns(typeConverter2, patterns2);
-
-  if (failed(
-          applyFullConversion(getOperation(), target2, std::move(patterns2)))) {
-    signalPassFailure();
-  }
-
-  // Lowering for nyazy.print -> LLVM printf
-  mlir::ConversionTarget printTarget(getContext());
-  printTarget.addLegalDialect<mlir::BuiltinDialect, mlir::LLVM::LLVMDialect>();
-  printTarget.addIllegalOp<nyacc::PrintOp>();
-
-  mlir::RewritePatternSet printPatterns(&getContext());
-  printPatterns.add<PrintIntOpLowering, PrintStringOpLowering>(&getContext());
-
-  if (failed(
-          applyFullConversion(getOperation(), printTarget, std::move(printPatterns)))) {
-    signalPassFailure();
-  }
-
 }
 
-std::unique_ptr<mlir::Pass> nyacc::createNyaZyToLLVMPass() {
-  return std::make_unique<NyaZyToLLVMPass>();
+std::unique_ptr<mlir::Pass> nyacc::createNyaZyToStdPass() {
+  return std::make_unique<NyaZyToStdPass>();
 }
